requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.5.0
django-extensions==3.0.9
django-filter==2.4.0
djangorestframework==3.12.1
gunicorn==20.0.4
importlib-metadata==2.0.0
Markdown==3.3.3
Pillow==8.0.1
python-decouple==3.3
pytz==2020.1
sqlparse==0.4.1
whitenoise==5.2.0
zipp==3.4.0
